requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.4
Flask-Login==0.6.2
python-dotenv==1.0.0
stripe==6.0.0
Werkzeug==2.3.7
gunicorn